Fix barrier and invalidation of the fill_buffer readback
The old code had a barrier waiting for a different memory region that is being
filled, and same goes for the memory invalidation call that follows.
It resulted in the range of size=0 being requeste for invalidation, which
doesn't include actual bytes filled by cmdFillBuffer that is being tested.

Affects:
EQP-VK.api.fill_and_update_buffer.*

Change-Id: Id2d68dc65145b42ecb0f966c70d57db81e9acb4a
Components: Vulkan