Fix VulkanPlatform compiling and linking on OSX
There where 2 problems with the code:
  1. it broke upon the change to vk::Library interface,
since OSX is not yet CI tested
  2. it didn't link against OpenGL library for Default target,
which appears to be required for anything on OSX

The latter was always an issue, but one we could work around locally.
I believe the proposed upstream patch is clean enough to avoid our hacks.

Change-Id: I9b570e3700031d92367a815baf5b32f751e248b3
Components: Framework